# Speed up encoding for int64, float32/64, and bool