Before:
```
BenchmarkEncodeInt64	 5000000	       301 ns/op
BenchmarkEncodeFloat64	 2000000	       892 ns/op
BenchmarkEncodeBool	 5000000	       255 ns/op
```

After:
```
BenchmarkEncodeInt64	10000000	       128 ns/op
BenchmarkEncodeFloat64	 5000000	       266 ns/op
BenchmarkEncodeBool	10000000	       124 ns/op
```

Relatively minor performance wins, but I thought I'd see what others thought.